+++ src/main/java/org/vuetiful/DNS/domain/member/service/MemberService.java
@@ -5,5 +5,5 @@ import org.vuetiful.DNS.domain.member.dto.MemberDTO;
 public interface MemberService {
     public String signup(MemberDTO memberDTO);
     public String login(MemberDTO memberDTO);
-    public String updateMember(Long memberId, MemberDTO memberDTO);
+    public String updateMember(int memberId, MemberDTO memberDTO);
 }
